dpRetrieve can now handle a file of file_ids or URLs
Prior to this patch dpRetrieve only worked on the local file system.
Now you can give a file of URLs or a file of file_ids and curl will
be used to retrieve the data. The data files are gzipped if
required.

Requires a .cadcuser file containing CADC user credentials.