CADC submissions are now about 3 seconds faster per entry
In commits 82e636f277f0566df98f43c1d4d002d09e825147 and
0fe9a65a28719bb6146e3cad7f5b34ed7cba45fe we changed from
an SQL statement that used max() to a select statement
that returned every row and then kept the last entry.

We did this because Sybase truncates trailing zeroes from
binary columns (00001200 becomes 000012). The problem is that
this is an incredibly inefficient way to find the max value
because it gets slower every time a new row is added. I've
now reverted to the original method of using max() and
the return value is now padded with zeroes to the required
16 characters. This changes the two queries from about 2.5 seconds
to 1e-5 seconds.