jsasubmit handles subheaders better
An OMP::Info::Obs sets up the object so that every file
exists in a subheader. This means that for a 250 file observation
(not uncommon for SCUBA-2) there are 250 subheaders and each
one needs to be associated with a grouping key. Creating an
ORAC::Frame object from a hash takes a finit amount of
time (0.15 sec on scubadev) so creating 250 of them takes about
30 seconds.

Since we know that each subsystem will need to be associated
with the same grouping key and so executed with the same
recipe we don't need to test each file separately. This patch
first associates each file with a subsystem and then works
out the grouping key. This turns 250 ORAC::Frame creations into
2 for SCUBA-2. Much quicker.