jsasubmit now supports updating of existing recipes instances
If the supplied tag matches one that already exists the recipe
instance is now updated. The state is changed from E or Y back
to " " (otherwise nothing happens) and the file list is
updated (deleting or adding as appropriate).

The return value of create_recipe_instance has changed so that
if a recipe instance existed previously but was not modified
because it was in the wrong state a "-" will be prepended.