lib: Move ASN_TYPE override into correct location
When asn_id code was refactored the code to update the ASN_TYPE
FITS header to "project" was moved into a section of code that
would work regardless of whether the file was a group file or
an obs file. This meant that obs files were flagged as project
type and then forced back to obs but now with the wrong ASN_TYPE
header.