By default do not use survey in ASN_ID
jsasubmit does project queries for surveys and so it risks duplicate
ASN_IDs if the same survey observation is done in two projects. By
default we now only use the project ID.

It may make sense at some point to have a "survey" mode along
with "project" to indicate when the survey is required.