Support new database schema
CADC have changed the DP database schema in two ways:

 1. All binary values 0x000... have been replaced by bigints
 2. dp_file_input and dp_recipe_instance now have identity columns
    that automatically insert these bigint values.

Current issue is that DBD::Sybase does not seem to be able to
handle bigints on INSERT with placeholders so the INSERTS,
DELETEs and UPDATEs have to use explicit SQL in strings.

Should only be released when CADC switch their system.