Add support for tiled public products
This commit updates the JSA modules as necessary for jsawrapdr to
be able to process the products generated by the PICARD recipe
COADD_JSA_TILES, which is intended to be run in "public" mode
to generate the public co-added tiles.  Filenames are provisional
pending final agreement but follow the decision from the JSA meeting
of 4/22/14 that the tile number is to remain at the end of the file
name (attached to the "healpix" product name).

* Added the filenames currently used by COADD_JSA_TILES to the pattern
  matches used to recognise and dissect reduced filesnames.

* Added a new output pattern for converted FITS files based on the
  association ID.

* Prepend the tile number to the association ID header in public mode.

* Allow the ASN_ID header to be set to "public".

* Use the plain product instead of product ID for preview images
  when in public mode.  This makes sense under the JSA rule that,
  other than in obs mode, the observation ID is the association ID,
  because since the association ID contains the filter (850um / 450um)
  this means that the two wavelengths would go into two different
  composite observations in CAOM-2.  If this turns out to be
  unsatisfactory (i.e. it is desired that the two wavelengths would
  go in the same composite observation) then this feature should
  be restricted to non-SCUBA-2 data.