JSA::Headers::CADC: update ASN_IDs for public mode
The hyphen-separated ASN_ID component removal now carefully removes
components one at a time to get back to either an ASN_ID without
hyphens, or one starting with a known precious component.

The correct_asn_id subroutine for JSA "public" mode is now instrument
specific (requiring the new jsa:instrume and jsa:backend headers in
PNG files).  For SCUBA-2 we just throw away the old ASN_ID (which
was the filter wavelength) and write a new one consisting of the
instrument name "SCUBA2" and tile number.  For HARP and ACSIS we
prepend the "HARP-ACSIS" and the tile number.