Add interface module for local jsa_proc system
The new module JSA::Proc will contain routines for interacting
with the new JSA local processing system.  It currently contains
a dummy implementation of a subroutine add_jsa_proc_jobs
for adding jobs to the system.  This is called by jsahpxobs with
the jobs submitted to CADC and those to be run locally.
The CADC recipe instance identifiers are now added to the hash
by the JSA::Submission::submit_jobs subroutine to allow this
information to be added to the local system.