JSA::Files: add mode argument to can_send_to_cadc
can_send_to_cadc is called before the association type is updated
(e.g. from night to public).  This means that when it calls
_can_send_to_cadc_quick, it gives the wrong association type
for group products when not in obs or night modes.  It looks like
this can only be corrected by passing the mode to can_send_to_cadc
so that it can update the mode as read from the file when
necessary.