jsasubmit: no longer process EC project (apart from EC05)
After consultation with Per, it appears that no one generally
uses our night processing of the EC projects. They are the cause
of a lot of errors, as they are often running hand-crafted XML
or require not-yet-written DR features.  This change will stop
them being included in the nightly jobs.