Add option to specify file version
This allows the "version" part of the CADC file names to be specified.

Currently the version is only applied when the association type is
"public".  Since no other types currently have versions other than 0,
this allows us to assume that other "public" files in the provenance
are also of the same version while others are of version 0.  If other
types start having versions, or if "public" products are constructed
from other "public" products of different versions, then we will need
an improved method of determining their versions -- perhaps an additional
FITS header would be required.

As an example of how this works for HEALPix "public" products:

The HEALPix "obs" files are always version 0:
    jcmts20140627_00060_850_healpix000514_obs_000.fits

The HEALPix "public" (coadd) gets a specified version, e.g. 1:
    jcmts850um_healpix000514_pub_001.fits
    PRV1 = 'jcmts20140627_00060_850_healpix000514_obs_000'
    (non-"public" provenance file, assumes version 0)

The catalogs also get the specified version, e.g. 1:
    jcmts850um_extent-mask000514_pub_001.fits
    PRV1 = 'jcmts850um_healpix000514_pub_001'
    ("public" provenance file, assumes same version)