JSA::Files: no longer add blank tile beside rsp
Requested by SFG.

Since CADC's Advanced Search interface only shows preview images when
the preview link is hovered over or clicked and can mix heterodyne
observations and continuum observations (for which there are only
rimg previews that we don't have a blank tile to the left of),
the blank tile to the right of lone rsp images doesn't help to align
the images on the page.

In addition, apparently some users of the archive have been confused
by the blank tiles as they can be interpreted as missing information
due to some kind of processing error.