JSA::EnterData::StarCommand: simplify module
Removed unnecessary object-orientation.

Combine running a command and retrieving its parameters into
one function.  This allows a temporary ADAM directory to be
selected so that we can be sure we retrieve the parameters from
the correct instance of the command.

Updated JSA::EnterData::calc_radec to actually use values retrieved
by this module rather than having a separate set of code to run PARGET.

Now check the return status from PARGET and return values as undef if
it failed to retrieve them.