JSA::EnterData: combine insert_hash methods
This combines the 'insert_hash' and 'conditional_insert_hash' methods.
When performing a "conditional" insert, now pre-filter the list of
data to be inserted to remove entries already present.  This allows
us to use placeholder values, eliminating a lot of complexity.  This
could perhaps be updated later to use a MySQL specific query such
as INSERT ... ON DUPLICATE KEY ...