JSA::EnterData: update transform_value for MySQL
Since "%transform_data" only appears to be used for boolean types,
and since the series of if-blocks deals with other types, replace
it with a dedicated "%transform_bool" hash.  This is then checked
if the data type starts with "int" or "tinyint" (to allow for a
following size -- the alternative would be to remove the size
specification first and then check the exact type name).