JSA::EnterData: add new method to combine INBEAM
To work properly, I think the INBEAM logic needs to be applied both
when examining a set of headers we are reading from files, and when
combining merging the current database value with an new one.
Therefore a new method has been created to capture this logic,
so that it can be applied in both places.  A new test script has
been added to check that it is working as desired.