JSA::EnterData: handle undef when combining INBEAM
Now handle undef (to mean nothing present) in the new INBEAM combining
method, rather than in the prepare_update_hash method for the case of
updates to existing records.  This is so that we set the database column
to NULL when there nothing in the beam during the initial insert, as well
as during updates.