JSA::EnterData: set placeholder for null INBEAM
Since we use the OMP classes to merge FITS headers and these
in turn use Astro::FITS::Header which is discarding null/blank
headers, we must set a placeholder for null INBEAM values in
order for them to be retained.  The placeholder is then
stripped out again when we process the INBEAM header.