JSA::EnterData: change calcbounds f/ns to methods
This changes some of the imported calcbounds routines from being
functions to being methods.  The reason for this is to allow them
to get the instrument object rather than having to have it passed
as an argument.

(Note: this adds a requirement that only one instrument be in the
list when using these calcbounds routines, as we just get the first.)