JSA::EnterData: move touched files hash to object
This commit moves the hash of touched files from being a global variable
in the JSA::EnterData module to being an attribute of each instance of
the JSA::EnterData class.  (This will prevent multiple EnterData
objects from clearing each other's cache.)

Note: it is not clear that this facility is ever useful.  The available
logs do not show any cases of observations being rejected because the
files were already touched.  The corresponding log message has been
upgraded from "trace" to "debug" level so that it is more prominent
in future.  The hash has been moved into the object as part of an attempt
to rationalize how this module handles information, but perhaps it
should just be removed entirely.