JSA::EnterData: use table names in %columns
Now use the instrument table name rather than the instrument name
in the "columns" hash.  This simplifies constructing the hash as
it simply maps tables to their columns listings and it prevents
us from having to correct the table name for SCUBA-2 (since it
is the instrument where the table name doesn't match the instrument
name).