JSA::DB: move DB handle caching to this module
Database handle caching is now performed by the JSA::DB module rather
than the JSA::DB::MySQL module.  This is because this functionality
is not database-specific and the JSA::DB::MySQL::connect_to_db function
was misleading in that it didn't always connect to a database.  JSA::DB
now performs two levels of caching: by object instance and then on a
module-wide basis in order to replicate the previous behavior.

This also allows us to check that the database connection is still
alive (using the $dbh->ping() method) and reconnect if not.