JSA::DB::TableTransfer: take a JSA::DB object
Since the software which uses this class connects to the database
using JSA::DB and since this class often wants a JSA::DB object,
pass the JSA::DB object to its constructor rather than the database
handle obtained from that object.  This avoids it from having to
construct new JSA::DB objects when one is needed.  It also means
that we can use JSA::DB's dbhandle method, which will reconnect
to the database if needed.