JSA::EnterData: simplify fill_max_subscan method
This method bizarrely used to fill in the "max_subscan" value with
the number of files rather than the maximum subscan number.

It has been updated to simply take the maximum value of the NSUBSCAN
header.  It is no longer instrument-specific.  This avoids the need to
connect to the database to get a file listing (in the case of SCUBA-2).
The prepare_update_hash method now keeps the maximum value of this
parameter.