JSA::EnterData: check for nsubscan values directly
If there were more than one file, the fill_headers_FILES method would
always look at the subheaders to get the subscan numbers.  This does
not work in the case of multiple files for the same subscan because
the headers will have been merged so that NSUBSCAN is not present in
the subheaders.  Now this method directly checks the status of the
header and should act appropriately.

(Prior to commit 9a2d0cc8e3 the SCUBA-2 class used to detect when the
number of files and nsubscan values didn't match and apply a correction
based on file name patterns.  That was removed as fill_headers_FILES
is supposed to perform this task -- it's better for it to be done
correctly the first time rather than having to duplicate the work for
the problematic case.)