JSA::DB::MongoDB: support reading from FITS files
In order to allow RxH3 raw data to be processed, this commit adds
support for reading headers from FITS files.  It also adds an extra
entry to the MongoDB record for the "subheaders" since the RxH3 raw
data has a lot of useful information in the header of the first
extension.