JSA::EnterData::RxH3: infer OBSNUM more thoroughly
When counting files to construct an OBSNUM header for RxH3 data files,
now check the raw data directories for the UT and HST date of interest.
This is because sometimes RxH3 data are located in a directory named
for the night's UT date and sometimes by the night's HST date (depending
on whether RxH3 observing was started before or after midnight).