JSA::EnterData: Separate statements in update_hash
The update_hash method now prepares a new update statement for each
of the update operations it is given.  This is because prepare_update_hash
does not guarantee that all updates will have the same set of columns
to be changed.  Typically there would only be one update so this
may not have caused a problem so far.

(It would also have returned after doing the first update, so any
additional updates specified would not have been done.)