JSA::EnterData: correct _apply_kludge_for_COMMON
This method checked for its argument being an arrayref, or having
more than 1 entry -- it should have checked for both of these
(using an "and" expression).  Also, in the case of returning early,
it should have returned the input reference, since "_update_or_insert"
always overwrites its list of headers with the value returned by this
method when processing inserts for the COMMON table.

(I'm not sure if this method is actually ever necessary any more.
It's certainly not necessary to group by "obsid" since we process
one observation at a time.  At least it now actually does what it
looks like it should be doing.  Also if the argument wasn't an
arrayref the second test (done with "or") would have caused an
error -- with "and" it won't be reached.)