JSA::EnterData: limit columns in get_insert_values
The restrictions on which columns should be updated are now applied
in the "get_insert_values" method, rather than "prepare_update_hash".
The helper method "extract_column_headers" has been merged into
"get_insert_values".

I think this is a more logical place to applying the filtering.  It
also simplifies "prepare_update_hash" and allows for better debugging
as we now know exactly what data we are trying to put into the
database before we start comparing it to the existing values.