JSA::EnterData: control range headers via dates
The updating of the range headers should now be controlled via the
DATE-OBS and DATE-END headers -- the start values will be updated if
DATE-OBS <= the database value and the end values will be updated if
DATE-END >= the database value.  Hopefully this will give the correct
behavior for incremental updates.

The main loop of the "prepare_update_hash" method has been changed
substantially.  It now loops over values being set, rather than
values already in the database, as in some modes we are only interested
in a subset of the values.  Then the check for the values being the
same as those currently in the database has been combined near the
top of the loop to separate it from the different types of updates
(range, maximum or normal).

Various helper functions which are no longer required have been removed.