JSA::EnterData: prepare_update_hash overwrite mode
This adds an "overwrite" mode option to the prepare_update_hash
method.  Setting this to a true option requests that database
values be updated simply rather than merged using the normal
rules for incremental data entry.

The new parameter is passed between the relevant methods.  It is
enabled directly in calcbounds_update_bound_cols (although the
bounds columns don't have merging rules so this isn't important).