JSA::EnterData: prepare common & inst data togther
Now prepare data structures containing the new and existing database
values in the "insert_observation" and "calcbounds_update_bound_cols"
methods.  These are now passed to the "_update_or_insert" method
which passes them on to the "prepare_update_hash" method.  The routine
to fetch existing database values has been extracted into a new
method "_get_existing_values".

The reason for this is so that we gather the information for the
COMMON and instrument-specific tables together.  This will allow
us to perform updates which require information from both tables.