Update OMP database usage
The prepare_archive_db method now returns an OMP::DB::Archive
object which should be used to perform queries.  The find_observations
method now requires this as its first argument.  In addition,
if it is given a project identifier, and not told not to
verify it, it requires an OMP::DB::Backend object via
a new DB optional argument.