Convert: include prov. in convert_dr_files dry run
The convert_dr_files function now passes on the dry_run option
to JSA::Starlink::prov_update_parent_path, which has been updated
to not write the modified provenance back to the file in this case.
The name of any file being skipped due to an error in provenance
processing is now included in the corresponding error message.