Convert: avoid error if prov. input cannot be read
Check that a file exists and also make failure to read the header
non-fatal.  (Retaining both approaches in order to give more useful
log output when $DEBUG is enabled.)

This is to allow for cases where a group product was listed in the
provenance of observation products, but then the pipeline's QA checks
caused the group product to be removed.